requirements.txt
flask
tensorflow
keras
numpy
pillow
utils
transformers
Pillow
